Add support for ESP32 based boards
The Wemos C3 Mini/Pico family are ESP32-C3 based boards that are pin
compatible with the Wemos D1 series, and as such fit on the Airgradient
PCBs. The pin names for the serial ports used to communicate with the
CO2 sensor and PMS sensor are different, though. Also, the ESP32-C3
supports more hardware UARTs.

This patch sets the correct pin names based on the microcontroller family.

It also adds support for using additional hardware RS232 modules that
are available on ESP32. This is controlled by the CO2_SERIAL_HARDWARE
and PMS_SERIAL_HARDWARE defines. Setting them to -1 indicates
SoftwareSerial, any number >= 0 indicates the hardware RS232 UART to
use.

The pin-compatible Wemos C3 Mini only has one available hardware UART
(the first is used for the serial console). By default this is assigned
to the CO2 sensor, with PMS using SoftwareSerial.

On an ESP8266 board both CO2 and PMS are using SoftwareSerial, as no
hardware UARTs are available.

Aside from this change, the library works fine on an ESP32.

x